restructure addon folder, move most things into addon/-private
