+++ addon/-private/helpers.js
