+++ addon/helpers/perform.js
@@ -1,5 +1,5 @@
 import { helper } from '@ember/component/helper';
-import { taskHelperClosure } from 'ember-concurrency/-helpers';
+import { taskHelperClosure } from 'ember-concurrency/-private/helpers';
 
 export function performHelper(args, hash) {
   return taskHelperClosure('perform', 'perform', args, hash);
