--- app/initializers/ember-concurrency.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-concurrency/initializers/ember-concurrency';
