+++ package.json
@@ -69,6 +69,7 @@
   ],
   "dependencies": {
     "babel-core": "^6.24.1",
+    "broccoli-file-creator": "^2.1.1",
     "ember-cli-babel": "^6.8.2",
     "ember-compatibility-helpers": "^1.2.0",
     "ember-maybe-import-regenerator": "^0.1.5"
